make a few more samples runnable
