+++ lib/execution.dart
@@ -102,6 +102,8 @@ class ExecutionContect {
   }
 
   void block(Bytecode code) {
+    // todo:
+
     throw 'unimplemented: block';
   }
